applicationcontroller7
Application controller for basic level. This is a simple application controller using spring framework and maven pom.xml. Second submission, first returned because it did not include a hashmap. This one is designed to add messages by number to a hashmap. Not a complete program, but the start using mvc and the application controller pattern.